move useDashboardResults into dashboardActions
