+++ public/app/features/commandPalette/actions/dashboardActions.ts
@@ -1,3 +1,6 @@
+import debounce from 'debounce-promise';
+import { useEffect, useState } from 'react';
+
 import { locationUtil } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
 import { t } from 'app/core/internationalization';
@@ -9,6 +12,8 @@ import { RECENT_DASHBOARDS_PRORITY, SEARCH_RESULTS_PRORITY } from '../values';
 
 const MAX_SEARCH_RESULTS = 100;
 const MAX_RECENT_DASHBOARDS = 5;
+
+const debouncedDashboardSearch = debounce(getDashboardSearchResultActions, 200);
 
 export async function getRecentDashboardActions(): Promise<CommandPaletteAction[]> {
   const recentUids = (await impressionSrv.getDashboardOpened()).slice(0, MAX_RECENT_DASHBOARDS);
@@ -68,4 +73,19 @@ export async function getDashboardSearchResultActions(searchQuery: string): Prom
   });
 
   return goToDashboardActions;
+}
+
+export function useDashboardResults(searchQuery: string, isShowing: boolean) {
+  const [dashboardResults, setDashboardResults] = useState<CommandPaletteAction[]>([]);
+
+  // Hit dashboards API
+  useEffect(() => {
+    if (isShowing && searchQuery.length > 0) {
+      debouncedDashboardSearch(searchQuery).then((resultActions) => {
+        setDashboardResults(resultActions);
+      });
+    }
+  }, [isShowing, searchQuery]);
+
+  return dashboardResults;
 }
